refactor root app to not load everything when visiting landing page
# Conflicts:
#	website/client/src/app.vue
